# api-eventos/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Generación y Validación de documentos de participación",
        "description": "Listado de rutas de la API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/events": {
            "get": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Listado de todos los registros de los eventos.",
                "operationId": "7058ab7c89d20c752b6afd6d808f6526",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "startDate": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-05-18 09:00:00"
                                        },
                                        "endDate": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-05-19 17:00:00"
                                        },
                                        "nameEvent": {
                                            "type": "string",
                                            "example": "Academia Journals"
                                        },
                                        "description": {
                                            "type": "text",
                                            "example": "Espacio que da a conocer resultados de investigación"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Eventos"
                ],
                "summary": "Almacena un recurso creado.",
                "operationId": "508895e03d7f10edac84357084c87b20",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "startDate",
                                    "endDate",
                                    "nameEvent",
                                    "description"
                                ],
                                "properties": {
                                    "startDate": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-05-28 09:00:00"
                                    },
                                    "endDate": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-05-29 19:00:00"
                                    },
                                    "nameEvent": {
                                        "type": "string",
                                        "example": "Academia Journals"
                                    },
                                    "description": {
                                        "type": "text",
                                        "example": "Espacio de trabajo que da a conocer los resultados de trabajos de investigación."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Evento registrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error en la solicitud",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Faltan campos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}